Projects/meal-planner/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

SMTP_HOST = os.getenv("SMTP_HOST", "smtp.gmail.com")
SMTP_PORT = int(os.getenv("SMTP_PORT", 587))
SMTP_USER = os.getenv("SMTP_USER")
SMTP_PASSWORD = os.getenv("SMTP_PASSWORD")
EMAIL_TO = os.getenv("EMAIL_TO")

HOUSEHOLD = {
    "adults": [
        {
            "label": "Mann",
            "height_cm": 191,
            "activity": "aktiv",
            "notes": "Hoher Proteinbedarf, sättigende Mahlzeiten",
        },
        {
            "label": "Schwangere Frau",
            "height_cm": 177,
            "activity": "moderat",
            "notes": "Nährstoffdicht, wenig leere Kohlenhydrate, schwangerschaftssicher",
            "restrictions": ["kein rohes Fisch", "keine unpasteurisierten Produkte", "kein Weichkäse"],
        },
    ],
    "children": [
        {
            "label": "Kind",
            "age_years": 3,
            "notes": "Manchmal wählerisch, kinderfreundliche Anpassungen ok",
        }
    ],
    "portion_multiplier": 2.5,  # 2 adults + 1 toddler
}

DIETARY_RULES = {
    "primarily_vegetarian": True,
    "allowed_proteins": ["Hähnchen", "Pute", "Fisch (kein Schalentier, kein rohes Fisch)"],
    "forbidden": ["Rindfleisch", "Schweinefleisch", "Schalentiere", "roher Fisch"],
    "high_protein_sources": ["Eier", "Hülsenfrüchte", "Tofu", "Tempeh", "Griechischer Joghurt", "Hüttenkäse", "Hähnchen", "Weißfisch"],
    "max_prep_minutes": 30,
    "budget_conscious": True,
    "stores": ["Lidl", "Rewe"],
}

RECIPE_SOURCES = [
    "https://www.chefkoch.de",
    "https://www.lecker.de",
    "https://eatsmarter.de",
    "https://www.yazio.com/de/gesunde-rezepte",
    "https://www.kuechengoetter.de",
    "https://www.kochbar.de",
]

CUISINE_ROTATION = ["Mediterran", "Asiatisch", "Deutsch", "Orientalisch", "Mexikanisch", "Französisch", "Griechisch"]

SYSTEM_PROMPT = """
Du bist ein Familien-Mahlzeitenplaner. Du erhältst eine Liste von Rezepten von deutschen Rezept-Websites
und erstellst einen 7-Tage-Mahlzeitenplan für einen bestimmten Haushalt.

Haushalt:
- Erwachsener Mann (groß, aktiv, braucht viel Protein und sättigende Mahlzeiten)
- Schwangere Frau (möchte Körperfett reduzieren, braucht nährstoffdichte, schwangerschaftssichere Ernährung —
  kein roher Fisch, keine unpasteurisierten Milchprodukte, kein Weichkäse)
- 3-jähriges Kind (manchmal wählerisch, braucht ggf. einfache Anpassungen)

Regeln:
- Vorwiegend vegetarisch, Geflügel und Fisch (kein Schalentier) erlaubt
- Max. 30 Minuten Zubereitung pro Rezept
- Keine wiederholten Mahlzeiten in derselben Woche
- Alle Zutaten bei Lidl oder Rewe in Deutschland erhältlich
- Budgetbewusst
- Hoher Protein- und Gemüseanteil
- Abwechselnde Küchenstile (Mediterran, Asiatisch, Deutsch, Orientalisch usw.)

Pro Tag werden GENAU 4 verschiedene Rezepte benötigt:
- lunch_option_1: einzigartiges Mittagessen nur für diesen Tag
- lunch_option_2: anderes einzigartiges Mittagessen nur für diesen Tag
- dinner_option_1: einzigartiges Abendessen nur für diesen Tag
- dinner_option_2: anderes einzigartiges Abendessen nur für diesen Tag

KRITISCHE REGEL: Über die gesamte Woche (alle 7 Tage × 4 Slots = 28 Slots) darf
kein Rezept mehr als EINMAL erscheinen. Jeder Slot braucht ein eigenes, neues Rezept.
Prüfe vor der Ausgabe: Sind alle 28 Rezeptnamen eindeutig verschieden?

Pro Rezept:
- Name und Quell-URL
- Zutaten mit metrischen Mengenangaben (g, ml, EL, TL) — skaliert für 2.5 Portionen (2 Erwachsene + 1 Kleinkind)
- Zubereitungsschritte
- Zubereitungszeit
- Kalorien und Protein pro Erwachsenenportion
- Kinderanpassung falls nötig (z.B. "Für das Kind: ohne Chili, Nudeln weicher kochen.")
- Schwangerschaftssicherheit ausdrücklich bestätigt

Danach 2 Einkaufslisten erstellen:
- Liste 1 (shopping_list_1): enthält ALLE Zutaten aller Rezepte die in lunch_option_1, lunch_option_2, dinner_option_1, dinner_option_2 von Montag, Dienstag, Mittwoch und Donnerstag stehen.
- Liste 2 (shopping_list_2): enthält ALLE Zutaten aller Rezepte die in lunch_option_1, lunch_option_2, dinner_option_1, dinner_option_2 von Freitag, Samstag und Sonntag stehen.

KRITISCH: Gehe jeden Tag einzeln durch. Schlage für jeden Tag alle 4 Rezepte nach und füge JEDE Zutat in die richtige Liste ein. Keine Zutat darf fehlen. Eine Zutat die in einem Rezept steht das an einem Montag–Donnerstag-Tag vorkommt, MUSS in Liste 1 sein — auch wenn das gleiche Rezept später in der Woche nochmal erscheint (dann auch in Liste 2).

Gruppierung nach: Gemüse & Obst / Proteine / Milch & Eier / Trockenwaren & Vorräte / Tiefkühl.
Mengen über alle betroffenen Rezepte konsolidieren.
Bei teuren Zutaten günstige Alternativen vorschlagen.
Wo sinnvoll: notieren ob Artikel eher bei Lidl oder Rewe erhältlich.

Ausgabe zuerst als strukturiertes JSON mit folgenden Keys:
- meal_plan: Array mit 7 Tages-Objekten (day, lunch_option_1, lunch_option_2, dinner_option_1, dinner_option_2)
- recipes: Array aller Rezept-Objekte
- shopping_list_1: Montag–Mittwoch
- shopping_list_2: Donnerstag–Sonntag

Die gesamte Ausgabe muss auf Deutsch sein.
"""
